requirements.txt
discord.py-self
requests
colorama
python-dateutil
decopt
pystyle